chore: refactor and break down functions to utils
