+++ client/src/components/ContentNavigator/ContentDataProvider.ts
@@ -674,18 +674,17 @@ class ContentDataProvider
     }
 
     const newUri = await this.model.moveTo(item, targetUri);
-    if (Array.isArray(closedFiles) && closedFiles.length > 0) {
-      // Reopen only the files that were closed
-      for (const closedFileUri of closedFiles) {
-        // Calculate the new URI for each closed file using the adapter
-        if (typeof newUri === "object" && newUri) {
-          const newFileUri = this.model
-            .getAdapter()
-            .calculateNewFileUri?.(closedFileUri, item, newUri);
-          if (newFileUri) {
-            await commands.executeCommand("vscode.open", newFileUri);
-          }
-        }
+    if (!newUri || !Array.isArray(closedFiles)) {
+      return !!newUri;
+    }
+    // Reopen only the files that were closed
+    for (const closedFileUri of closedFiles) {
+      // Calculate the new URI for each closed file using the adapter
+      const newFileUri = this.model
+        .getAdapter()
+        .calculateNewFileUri?.(closedFileUri, item, newUri);
+      if (newFileUri) {
+        await commands.executeCommand("vscode.open", newFileUri);
       }
     }
